seperated then ingestion script and sql stuff
